Fix some bugs and added Not found page and resetform function for booking form.
