+++ src/components/ui/card.jsx
@@ -10,7 +10,8 @@ function Card({
     <div
       data-slot="card"
       className={cn(
-        "bg-card text-card-foreground flex flex-col gap-6 rounded-xl border py-6 shadow-sm",
+        "bg-white/70 backdrop-blur-md text-card-foreground flex flex-col gap-6 rounded-xl border border-white/20 py-6 shadow-lg",
+        "hover:bg-white/80 transition-all duration-300",
         className
       )}
       {...props} />
@@ -26,6 +27,7 @@ function CardHeader({
       data-slot="card-header"
       className={cn(
         "@container/card-header grid auto-rows-min grid-rows-[auto_auto] items-start gap-1.5 px-6 has-data-[slot=card-action]:grid-cols-[1fr_auto] [.border-b]:pb-6",
+        "border-b border-white/10",
         className
       )}
       {...props} />
@@ -39,7 +41,7 @@ function CardTitle({
   return (
     <div
       data-slot="card-title"
-      className={cn("leading-none font-semibold", className)}
+      className={cn("leading-none font-semibold text-gray-800", className)}
       {...props} />
   );
 }
@@ -51,7 +53,7 @@ function CardDescription({
   return (
     <div
       data-slot="card-description"
-      className={cn("text-muted-foreground text-sm", className)}
+      className={cn("text-gray-600 text-sm", className)}
       {...props} />
   );
 }
@@ -75,7 +77,13 @@ function CardContent({
   className,
   ...props
 }) {
-  return (<div data-slot="card-content" className={cn("px-6", className)} {...props} />);
+  return (
+    <div 
+      data-slot="card-content" 
+      className={cn("px-6", className)} 
+      {...props} 
+    />
+  );
 }
 
 function CardFooter({
@@ -85,7 +93,11 @@ function CardFooter({
   return (
     <div
       data-slot="card-footer"
-      className={cn("flex items-center px-6 [.border-t]:pt-6", className)}
+      className={cn(
+        "flex items-center px-6 [.border-t]:pt-6",
+        "border-t border-white/10",
+        className
+      )}
       {...props} />
   );
 }
