+++ src/index.css
@@ -3,6 +3,24 @@
 
 @custom-variant dark (&:is(.dark *));
 
+/* Gradient Animation */
+@keyframes gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+
+.animate-gradient {
+  background-size: 200% 200%;
+  animation: gradient 15s ease infinite;
+}
+
 :root {
   --radius: 0.625rem;
   --background: oklch(1 0 0);
