+++ src/protectedRoute.jsx
@@ -2,8 +2,6 @@ import { Navigate } from "react-router-dom";
 
 const ProtectedRoute = ({ element }) => {
     const isAuthenticated = localStorage.getItem("token");
-    console.log(isAuthenticated);
-
     return isAuthenticated ? element : <Navigate to="/" />;
 };
